requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
defusedxml==0.7.1
dill==0.3.7
discord.py==2.3.2
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
djangorestframework==3.14.0
filelock==3.12.4
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
mysqlclient==2.2.0
packaging==23.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2-binary==2.9.9
pylint==3.0.1
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.5
whitenoise==6.6.0
yarl==1.9.2
